+++ stack/themes/mrn-base-stack/template-parts/content-case_study.php
@@ -236,7 +236,7 @@ $mrn_tags_list       = get_the_term_list( $mrn_post_id, 'post_tag', '', esc_html
 
 			<?php if ( has_post_thumbnail() ) : ?>
 				<a class="post-thumbnail" href="<?php the_permalink(); ?>" aria-hidden="true" tabindex="-1">
-					<?php the_post_thumbnail( 'large' ); ?>
+					<?php echo mrn_base_stack_get_lazy_post_thumbnail( null, 'large' ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped ?>
 				</a>
 			<?php endif; ?>